refactor(ir): isolate imported call planning
Move source-unit and ambient-host imported-call planning into a focused module so the new class-member lowering path stays within source and function budgets. Preserve exact owner and target identity checks while keeping ordinary module imports distinct from ambient env imports.
