+++ src/codegen/index.ts
@@ -26,8 +26,6 @@ import {
 import type { FieldDef, Instr, StructTypeDef, ValType, WasmFunction, WasmModule } from "../ir/types.js";
 import { createEmptyModule } from "../ir/types.js";
 import { compileIrPathFunctions, type IrIntegrationError, type IrIntegrationReport } from "../ir/integration.js";
-import { irArgcGlobalRef, irSupportGlobalRef } from "../ir/abi-bindings.js";
-import { irImportFuncRef, irSupportFuncRef } from "../ir/callable-bindings.js";
 import { asVal, irDynamic, isDynamic, irVal, type IrFuncRef, type IrType } from "../ir/nodes.js";
 import type { LatticeType } from "../ir/propagate.js";
 import {
@@ -38,7 +36,6 @@ import {
   type IrPreparationFailure,
 } from "../ir/outcomes.js";
 import {
-  certifyImportedIrCall,
   effectiveIrParamTypeNode,
   effectiveIrReturnTypeNode,
   irClosureSignatureFromFunctionTypeNode,
@@ -48,14 +45,12 @@ import {
 import type {
   IrHostVoidCallbackLoweringPlan,
   IrImportedCallLoweringPlan,
-  IrImportedOptionalParamPlan,
   IrTopLevelFunctionValueLoweringPlan,
 } from "../ir/ast-lowering-plans.js";
 import {
   makeIrAmbientClassCallResolver,
   makeIrHostGlobalResolver,
   makeIrHostVoidCallbackResolver,
-  type IrAmbientClassCallResolver,
 } from "../ir/host-extern.js"; // (#2856/#3214/#3657)
 import { makeIrHostDateSnapshotResolver } from "../ir/host-date.js";
 import { supportsIrBackendTargetCapability, type IrBackendTargetCapability } from "../ir/backend/legality.js";
@@ -91,6 +86,11 @@ import { createCodegenContext } from "./context/create-context.js";
 import { ProgramAbiSession, type PublishedProgramAbi } from "./program-abi-session.js";
 import { planProgramAbiFunctionValue, planProgramAbiGlobal, PROGRAM_ABI_GLOBAL_ROLE } from "./program-abi-planning.js";
 import { collectLocalCallEdgesByIdentity } from "./ir-first-gate.js";
+import {
+  planIrImportedCalls,
+  prepareIrAmbientClassCallLowering,
+  recordIrOverlayPreparationFailure,
+} from "./ir-imported-call-planning.js";
 import {
   applyIrFinalContextFunctionRetention,
   closeIrBlockedComponentByIdentity,
@@ -1642,23 +1642,6 @@ interface IrOverlayPlan {
   readonly importedFunctionResolver?: irOverlayIdentity.IrIdentityImportedFunctionResolver;
 }
 
-function recordIrOverlayPreparationFailure(
-  plan: Pick<IrOverlayPlan, "identityPlan" | "preparationFailuresByUnitId">,
-  legacyName: string,
-  failure: IrPreparationFailure,
-): void {
-  const unitId = irOverlayIdentity.requireIrOverlayUnitId(plan.identityPlan, legacyName);
-  const previous = plan.preparationFailuresByUnitId.get(unitId);
-  if (previous && previous !== failure) {
-    throw new IrInvariantError(
-      "selection-preparation-mismatch",
-      "resolve",
-      `IR unit ${unitId} / ${legacyName} received more than one preparation result`,
-    );
-  }
-  plan.preparationFailuresByUnitId.set(unitId, failure);
-}
-
 function synchronizeIrSafeFunctionSelection(plan: IrOverlayPlan, selection: IrSelection): IrSelection {
   const retainedUnitIds = new Set<IrUnitId>();
   for (const legacyName of selection.funcs) {
@@ -1754,88 +1737,6 @@ function prepareHostDateSnapshotPreflight(
   return retainedSelection;
 }
 
-function ambientHostValType(type: IrType | null): ValType | null | undefined {
-  if (type === null) return null;
-  if (type.kind === "string") return { kind: "externref" };
-  if (type.kind !== "val") return undefined;
-  if (type.val.kind === "f64" || type.val.kind === "i32") return type.val;
-  return undefined;
-}
-
-function sameAmbientHostValType(left: ValType, right: ValType): boolean {
-  if (left.kind !== right.kind) return false;
-  if ((left.kind === "ref" || left.kind === "ref_null") && (right.kind === "ref" || right.kind === "ref_null")) {
-    return left.typeIdx === right.typeIdx;
-  }
-  return true;
-}
-
-function hasPreparedAmbientHostImport(ctx: CodegenContext, plan: IrImportedCallLoweringPlan): boolean {
-  if (plan.source !== "ambient-host" || plan.target.binding.kind !== "import" || plan.target.binding.module !== "env") {
-    return false;
-  }
-  const params = plan.params.map(ambientHostValType);
-  const result = ambientHostValType(plan.returnType);
-  if (params.some((param) => param === undefined) || result === undefined) return false;
-
-  const targetName = plan.target.binding.field;
-  const funcIdx = ctx.funcMap.get(targetName);
-  if (funcIdx === undefined || funcIdx < 0 || funcIdx >= ctx.numImportFuncs) return false;
-  let importFuncIdx = 0;
-  for (const imported of ctx.mod.imports) {
-    if (imported.desc.kind !== "func") continue;
-    if (importFuncIdx++ !== funcIdx) continue;
-    if (imported.module !== plan.target.binding.module || imported.name !== targetName) return false;
-    const type = ctx.mod.types[imported.desc.typeIdx];
-    const expectedParams = params as ValType[];
-    const expectedResults = result === null ? [] : [result];
-    return (
-      type?.kind === "func" &&
-      type.params.length === expectedParams.length &&
-      type.results.length === expectedResults.length &&
-      type.params.every((param, index) => sameAmbientHostValType(param, expectedParams[index]!)) &&
-      type.results.every((actual, index) => sameAmbientHostValType(actual, expectedResults[index]!))
-    );
-  }
-  return false;
-}
-
-/**
- * #3657 — prove that declaration collection materialised every certified
- * class-member ambient call as the exact env function import before lowering.
- */
-function prepareAmbientClassCallLowering(
-  ctx: CodegenContext,
-  plan: IrOverlayPlan,
-  selection: IrSelection,
-): IrSelection {
-  const classMembers = new Set(selection.classMembers ?? []);
-  if (classMembers.size === 0) return selection;
-  const blocked = new Set<string>();
-  for (const callPlan of plan.importedCalls.values()) {
-    if (
-      callPlan.source === "ambient-host" &&
-      callPlan.target.binding.kind === "import" &&
-      callPlan.target.binding.module === "env" &&
-      classMembers.has(callPlan.ownerName) &&
-      !hasPreparedAmbientHostImport(ctx, callPlan)
-    ) {
-      blocked.add(callPlan.ownerName);
-    }
-  }
-  if (blocked.size === 0) return selection;
-  for (const ownerName of blocked) {
-    classMembers.delete(ownerName);
-    recordIrOverlayPreparationFailure(plan, ownerName, {
-      kind: "unsupported",
-      code: "late-preparation-unsupported",
-      stage: "resolve",
-      detail: "the checker-certified ambient function is absent from the final env import manifest",
-    });
-  }
-  return { ...selection, classMembers };
-}
-
 function recordWholeSourceFailure(
   ctx: CodegenContext,
   sourceFile: ts.SourceFile,
@@ -1873,271 +1774,6 @@ function recordObservedIrOutcomes(
   });
   ctx.irOutcomes.push(...reconciled.outcomes);
   for (const diagnostic of reconciled.diagnostics) reportErrorNoNode(ctx, diagnostic);
-}
-
-/**
- * A zero-result imported call is lowerable only where JavaScript discards its
- * value. Keep this check in the pre-claim planner: ordinary expression
- * lowering requires an SSA result and must never discover the missing value
- * after the legacy body has already been skipped.
- */
-function importedVoidCallIsDiscarded(call: ts.CallExpression, owner: ts.FunctionDeclaration): boolean {
-  let current: ts.Expression = call;
-  for (;;) {
-    const parent = current.parent;
-    if (ts.isParenthesizedExpression(parent) && parent.expression === current) {
-      current = parent;
-      continue;
-    }
-    // `void` always discards its operand, even when the resulting undefined
-    // value is itself consumed by an outer expression.
-    if (ts.isVoidExpression(parent) && parent.expression === current) return true;
-    // A conditional arm is discarded only when the conditional as a whole is
-    // discarded. Its condition is a value position and deliberately stops.
-    if (ts.isConditionalExpression(parent) && (parent.whenTrue === current || parent.whenFalse === current)) {
-      current = parent;
-      continue;
-    }
-    // Discard lowering evaluates comma operands in source order. Ascend to
-    // require the whole comma expression to reach a discarded context.
-    if (ts.isCommaListExpression(parent) && parent.elements.some((element) => element === current)) {
-      current = parent;
-      continue;
-    }
-    if (
-      ts.isBinaryExpression(parent) &&
-      parent.operatorToken.kind === ts.SyntaxKind.CommaToken &&
-      (parent.left === current || parent.right === current)
-    ) {
-      current = parent;
-      continue;
-    }
-    break;
-  }
-
-  const parent = current.parent;
-  if (ts.isExpressionStatement(parent) && parent.expression === current) return true;
-  return (
-    ts.isReturnStatement(parent) &&
-    parent.expression === current &&
-    effectiveIrReturnTypeNode(owner)?.kind === ts.SyntaxKind.VoidKeyword
-  );
-}
-
-interface IrImportedOverlayPlans {
-  readonly importedCalls: Map<ts.CallExpression, IrImportedCallLoweringPlan>;
-  readonly topLevelFunctionValues: Map<ts.Identifier, IrTopLevelFunctionValueLoweringPlan>;
-}
-
-function planIrAmbientClassCallLowering(
-  ctx: CodegenContext,
-  identityPlan: irOverlayIdentity.IrOverlayIdentityPlan,
-  resolveAmbientClassCall: IrAmbientClassCallResolver,
-  classShapeSidecar: IrClassShapeSidecar,
-  safeSelection: IrOverlayPlan["safeSelection"],
-  recordPreparationFailure: (legacyName: string, failure: IrPreparationFailure) => void,
-): Map<ts.CallExpression, IrImportedCallLoweringPlan> {
-  const importedCalls = new Map<ts.CallExpression, IrImportedCallLoweringPlan>();
-  const retainedClassMembers = new Set(safeSelection.classMembers ?? []);
-  if (retainedClassMembers.size === 0) return importedCalls;
-
-  for (const [ownerUnitId, owner] of identityPlan.identitySelection.classMembers ?? []) {
-    const ownerName = owner.legacyMatchName;
-    if (!retainedClassMembers.has(ownerName)) continue;
-    const declaration = identityPlan.identityContext.declarationByUnitId.get(ownerUnitId);
-    if (
-      !declaration ||
-      (!ts.isMethodDeclaration(declaration) &&
-        !ts.isGetAccessorDeclaration(declaration) &&
-        !ts.isSetAccessorDeclaration(declaration) &&
-        !ts.isConstructorDeclaration(declaration)) ||
-      !declaration.body
-    ) {
-      throw new IrInvariantError(
-        "selection-preparation-mismatch",
-        "resolve",
-        `ambient class-call owner ${ownerUnitId} / ${ownerName} has no exact executable declaration`,
-      );
-    }
-
-    let planningFailure: IrPreparationFailure | undefined;
-    const visit = (node: ts.Node): void => {
-      if (planningFailure) return;
-      if (node !== declaration && ts.isFunctionLike(node)) return;
-      if (ts.isCallExpression(node)) {
-        const certified = resolveAmbientClassCall(node);
-        if (certified) {
-          try {
-            const params = certified.declaration.parameters.map((parameter) =>
-              resolvePositionType(effectiveIrParamTypeNode(parameter), undefined, ctx, classShapeSidecar),
-            );
-            const returnType = resolvePositionType(
-              effectiveIrReturnTypeNode(certified.declaration),
-              undefined,
-              ctx,
-              classShapeSidecar,
-            );
-            importedCalls.set(node, {
-              source: "ambient-host",
-              ownerUnitId,
-              ownerName,
-              target: irImportFuncRef("env", certified.targetName),
-              params,
-              returnType,
-              optionalParams: new Map(),
-              needsArgc: false,
-            });
-          } catch (error) {
-            planningFailure = classifyIrFailure(error, "resolve");
-            return;
-          }
-        }
-      }
-      ts.forEachChild(node, visit);
-    };
-    visit(declaration.body);
-    if (planningFailure) {
-      recordPreparationFailure(ownerName, planningFailure);
-      retainedClassMembers.delete(ownerName);
-      for (const [call, plan] of importedCalls) {
-        if (plan.ownerUnitId === ownerUnitId) importedCalls.delete(call);
-      }
-    }
-  }
-  safeSelection.classMembers = retainedClassMembers;
-  return importedCalls;
-}
-
-function planIrImportedLowering(
-  ctx: CodegenContext,
-  identityPlan: irOverlayIdentity.IrOverlayIdentityPlan,
-  identityImportedFunctions: irOverlayIdentity.IrIdentityImportedFunctionResolver,
-  legacyImportedFunctions: ReturnType<typeof irOverlayIdentity.projectIrOverlayImportedResolver>,
-  classShapeSidecar: IrClassShapeSidecar,
-  declByName: ReadonlyMap<string, ts.FunctionDeclaration>,
-  safeSelection: IrOverlayPlan["safeSelection"],
-  recordPreparationFailure: (legacyName: string, failure: IrPreparationFailure) => void,
-): IrImportedOverlayPlans {
-  const importedCalls = new Map<ts.CallExpression, IrImportedCallLoweringPlan>();
-  const topLevelFunctionValues = new Map<ts.Identifier, IrTopLevelFunctionValueLoweringPlan>();
-  const planIdentity = irOverlayIdentity.makeIrFeaturePlanIdentity(identityPlan, identityImportedFunctions);
-  const entrySourceId = identityPlan.identityContext.inventory.sources.find((source) => source.kind === "entry")?.id;
-  if (!entrySourceId) {
-    throw new IrInvariantError(
-      "selection-preparation-mismatch",
-      "resolve",
-      "imported lowering requires one exact entry-source identity",
-    );
-  }
-  for (const [ownerName, declaration] of declByName) {
-    if (!safeSelection.funcs.has(ownerName) || !declaration.body) continue;
-    let planningFailure: IrPreparationFailure | undefined;
-    const visit = (node: ts.Node): void => {
-      if (planningFailure) return;
-      if (node !== declaration && ts.isFunctionLike(node)) return;
-      if (ts.isCallExpression(node)) {
-        const certified = certifyImportedIrCall(node, legacyImportedFunctions);
-        if (certified) {
-          try {
-            if (
-              process.env.JS2WASM_TEST_INJECT_IR_IMPORTED_PLAN_THROW === "1" ||
-              process.env.JS2WASM_TEST_INJECT_IR_IMPORTED_PLAN_THROW === ownerName
-            ) {
-              throw new Error(`injected imported-call planning failure for ${ownerName}`);
-            }
-            const params = certified.target.declaration.parameters.map((parameter) =>
-              resolvePositionType(effectiveIrParamTypeNode(parameter), undefined, ctx, classShapeSidecar),
-            );
-            const returnNode = effectiveIrReturnTypeNode(certified.target.declaration);
-            const returnType =
-              returnNode?.kind === ts.SyntaxKind.VoidKeyword
-                ? null
-                : resolvePositionType(returnNode, undefined, ctx, classShapeSidecar);
-            if (returnType === null && !importedVoidCallIsDiscarded(node, declaration)) {
-              throw new IrUnsupportedError(
-                "imported-call-planning-unsupported",
-                "resolve",
-                "void imported result is used in a value context",
-              );
-            }
-            if (returnType?.kind === "callable") {
-              throw new IrUnsupportedError(
-                "imported-call-planning-unsupported",
-                "resolve",
-                "callable imported results are outside A+B1",
-              );
-            }
-            const optionalParams = new Map<number, IrImportedOptionalParamPlan>();
-            for (const optional of ctx.funcOptionalParams.get(certified.target.targetName) ?? []) {
-              optionalParams.set(optional.index, {
-                ...(optional.constantDefault ? { constantDefault: optional.constantDefault } : {}),
-                ...(optional.hasExpressionDefault ? { hasExpressionDefault: true } : {}),
-              });
-            }
-            const importedIdentity = planIdentity.imported(ownerName, node.expression, certified.target);
-            const needsArgc =
-              ctx.funcUsesArguments.has(certified.target.targetName) ||
-              ctx.funcOptionalParams.has(certified.target.targetName);
-            importedCalls.set(node, {
-              source: "module-import",
-              ...importedIdentity,
-              ownerName,
-              params,
-              returnType,
-              optionalParams,
-              needsArgc,
-              ...(needsArgc ? { argcGlobal: irArgcGlobalRef(entrySourceId) } : {}),
-            });
-            for (const functionArgument of certified.functionArguments) {
-              const valueIdentity = planIdentity.value(ownerName, functionArgument.argument, functionArgument.target);
-              if (valueIdentity.target.binding.kind !== "unit") {
-                throw new IrInvariantError(
-                  "selection-preparation-mismatch",
-                  "resolve",
-                  `function-value target ${valueIdentity.target.name} has no exact source-unit binding`,
-                );
-              }
-              const trampolineName = `__fn_tramp_${functionArgument.target.targetName}_cached`;
-              const cacheGlobalName = `__fn_closure_${functionArgument.target.targetName}`;
-              topLevelFunctionValues.set(functionArgument.argument, {
-                ...valueIdentity,
-                ownerName,
-                signature: functionArgument.signature,
-                trampoline: irSupportFuncRef(
-                  valueIdentity.target.binding.unitId,
-                  "function-value-trampoline",
-                  trampolineName,
-                ),
-                cacheGlobal: irSupportGlobalRef(
-                  valueIdentity.target.binding.unitId,
-                  "function-value-cache",
-                  cacheGlobalName,
-                ),
-                cacheGlobalName,
-              });
-            }
-          } catch (error) {
-            planningFailure = classifyIrFailure(error, "resolve");
-            return;
-          }
-        }
-      }
-      ts.forEachChild(node, visit);
-    };
-    visit(declaration.body);
-    if (planningFailure) {
-      recordPreparationFailure(ownerName, planningFailure);
-      safeSelection.funcs.delete(ownerName);
-      irOverlayIdentity.dropIrSafeFunctionByLegacyName(identityPlan, ownerName);
-      for (const [call, plan] of importedCalls) {
-        if (plan.ownerName === ownerName) importedCalls.delete(call);
-      }
-      for (const [identifier, plan] of topLevelFunctionValues) {
-        if (plan.ownerName === ownerName) topLevelFunctionValues.delete(identifier);
-      }
-    }
-  }
-  return { importedCalls, topLevelFunctionValues };
 }
 
 function planIrOverlay(
@@ -2475,45 +2111,21 @@ function planIrOverlay(
     identityPlan.safeFunctionUnitIds,
     identityContext,
   );
-  let importedCalls = new Map<ts.CallExpression, IrImportedCallLoweringPlan>();
-  let topLevelFunctionValues = new Map<ts.Identifier, IrTopLevelFunctionValueLoweringPlan>();
+  const { importedCalls, topLevelFunctionValues } = planIrImportedCalls({
+    ctx,
+    identityPlan,
+    preparationFailuresByUnitId,
+    ...(jsHostExterns && identityImportedFunctions ? { identityImportedFunctions, legacyImportedFunctions } : {}),
+    ...(resolveAmbientClassCall ? { resolveAmbientClassCall } : {}),
+    classShapeSidecar,
+    safeSelection,
+    resolvePositionType: (node, mapped, classShapes) => resolvePositionType(node, mapped, ctx, classShapes),
+  });
   const hostVoidCallbacks = new Map<ts.ArrowFunction, IrHostVoidCallbackLoweringPlan>();
   const hostDateImportsByOwnerUnitId = new Map<
     IrUnitId,
     { ownerUnitId: IrUnitId; ownerName: string; importNames: Set<string> }
   >();
-  if (jsHostExterns && identityImportedFunctions) {
-    ({ importedCalls, topLevelFunctionValues } = planIrImportedLowering(
-      ctx,
-      identityPlan,
-      identityImportedFunctions,
-      legacyImportedFunctions,
-      classShapeSidecar,
-      declByName,
-      safeSelection,
-      recordPreparationFailure,
-    ));
-  }
-  if (resolveAmbientClassCall) {
-    const ambientCalls = planIrAmbientClassCallLowering(
-      ctx,
-      identityPlan,
-      resolveAmbientClassCall,
-      classShapeSidecar,
-      safeSelection,
-      recordPreparationFailure,
-    );
-    for (const [call, plan] of ambientCalls) {
-      if (importedCalls.has(call)) {
-        throw new IrInvariantError(
-          "selection-preparation-mismatch",
-          "resolve",
-          `ambient and source-unit imported-call plans overlap at ${ast.sourceFile.fileName}:${call.pos}`,
-        );
-      }
-      importedCalls.set(call, plan);
-    }
-  }
   if (resolveHostVoidCallback) {
     for (const [ownerName, declaration] of declByName) {
       if (!safeSelection.funcs.has(ownerName) || !declaration.body) continue;
@@ -3709,10 +3321,9 @@ export function generateModule(
       // classShapes → overrideMap → safeSelection → new.target gate).
       const plan = irPlan ?? planIrOverlay(ctx, ast, irPlanningIdentityContext!);
       const { classShapes, overrideMap } = plan;
-      let safeSelection = prepareAmbientClassCallLowering(ctx, plan, plan.safeSelection);
-      safeSelection = applyIrFinalContextFunctionUnitIds(
+      let safeSelection = applyIrFinalContextFunctionUnitIds(
         plan,
-        safeSelection,
+        prepareIrAmbientClassCallLowering(ctx, plan, plan.safeSelection),
         prepareHostVoidCallbackLoweringByIdentity(
           ctx,
           ast.sourceFile,
